Updated to add -mstrict-align and use Apple Silicon gcc

+++ part10-multicore/multicore.h
@@ -1,7 +1,7 @@
-extern unsigned int spin_cpu0;
-extern unsigned int spin_cpu1;
-extern unsigned int spin_cpu2;
-extern unsigned int spin_cpu3;
+extern unsigned long spin_cpu0;
+extern unsigned long spin_cpu1;
+extern unsigned long spin_cpu2;
+extern unsigned long spin_cpu3;
 
 void start_core1(void (*func)(void));
 void start_core2(void (*func)(void));
